поправил ошибки линтера
