+++ client/src/pages/LoginPage/LoginPage.tsx
@@ -1,9 +1,7 @@
 import { LoginForm } from "@/features/LoginForm";
 import classes from "./LoginPage.module.scss";
 
-interface LoginPageProps {}
-
-export const LoginPage = ({}: LoginPageProps) => {
+export const LoginPage = () => {
   return (
     <div className={classes.page}>
       <header></header>
