+++ client/src/shared/ui/AppLink/types.ts
@@ -1,0 +1,3 @@
+export enum AppLinkSize {
+   S = "s",
+}
