+++ client/src/shared/ui/Button/index.ts
@@ -1,1 +1,2 @@
-export { Button, ButtonSize } from './Button';
+export { Button } from "./Button";
+export { ButtonSize } from './types'
