+++ client/src/shared/ui/Button/types.ts
@@ -1,0 +1,3 @@
+export enum ButtonSize {
+   M = "m",
+}
