+++ client/src/shared/ui/Text/types.ts
@@ -1,0 +1,3 @@
+export enum TextSize {
+   S = "s",
+}
