+++ client/src/shared/ui/Title/types.ts
@@ -1,0 +1,3 @@
+export enum TitleSize {
+   XXL = "xxl",
+}
